# internal/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: CalBot API
  version: 0.1.0
  description: API for managing staff authentication and CalBot content.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/staff/prompts:
    get:
      tags: [Staff prompts]
      summary: List all prompt variants for a menu option
      description: Requires a STAFF bearer token. Active and inactive prompts are returned.
      security:
        - bearerAuth: []
      parameters:
        - name: menu_option_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: All prompts for the menu option
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffMenuOptionPrompt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Menu option not found
    post:
      tags: [Staff prompts]
      summary: Create a prompt variant for a menu option
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePromptRequest'
      responses:
        '201':
          description: Prompt created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MenuOptionPrompt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
  /api/v1/staff/response-groups:
    post:
      tags: [Staff response groups]
      summary: Create a response-group variant for a menu option
      description: A menu option can have multiple response groups. Each group represents an independent response variant and is saved with its items atomically.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResponseGroupRequest'
      responses:
        '201':
          description: Response group created with its items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '409':
          description: The menu option already has child options
  /api/v1/staff/response-groups/by-menu-option/{menuOptionID}:
    get:
      tags: [Staff response groups]
      summary: List all response groups and response items for a menu option
      description: Requires a STAFF bearer token. Active and inactive groups and items are returned.
      security:
        - bearerAuth: []
      parameters:
        - name: menuOptionID
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: All response groups and items for the menu option
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StaffResponseGroup'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Menu option not found
  /api/v1/staff/response-groups/{id}:
    delete:
      tags: [Staff response groups]
      summary: Permanently delete one response group and its items
      description: Requires a STAFF bearer token. Response items are deleted by cascade. Content feedback is independent and is not affected.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Response group and its items deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Response group not found
  /api/v1/staff/prompts/{id}:
    delete:
      tags: [Staff prompts]
      summary: Permanently delete a menu option prompt
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PromptID'
      responses:
        '204':
          description: Prompt deleted
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Prompt not found
  /api/v1/staff/prompts/{id}/active:
    patch:
      tags: [Staff prompts]
      summary: Set the active state of a menu option prompt
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PromptID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
      responses:
        '200':
          description: Prompt state updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          description: Missing or invalid staff token
        '403':
          description: Staff role required
        '404':
          description: Prompt not found
...
components:
  parameters:
    PromptID:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    MenuOptionPrompt:
      type: object
      required: [id, message]
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
    ResponseGroup:
      type: object
      required: [id, name, items]
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/ResponseItem'
    ResponseItem:
      type: object
      required: [id, type, sort_order]
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [TEXT, LINK, IMAGE, YOUTUBE, VIDEO, DOCUMENT]
        text:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        caption:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        sort_order:
          type: integer
...
    CreatePromptRequest:
      type: object
      required: [menu_option_id, message]
      properties:
        menu_option_id:
          type: integer
          format: int64
        message:
          type: string
        weight:
          type: integer
          minimum: 1
          default: 1
    CreateResponseGroupRequest:
      type: object
      required: [menu_option_id, name, items]
      properties:
        menu_option_id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        weight:
          type: integer
          minimum: 1
          default: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/ReplaceResponseItemRequest'
    ReplaceResponseItemRequest:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [TEXT, LINK, IMAGE, YOUTUBE, VIDEO, DOCUMENT]
        text:
          type: string
          nullable: true
          description: Required when type is TEXT.
        url:
          type: string
          nullable: true
          description: Required when type is LINK, IMAGE, YOUTUBE, VIDEO, or DOCUMENT.
        caption:
          type: string
          nullable: true
        metadata:
          type: object
          nullable: true
          additionalProperties: true
        sort_order:
          type: integer
          default: 0
...
    StaffMenuOptionPrompt:
      allOf:
        - $ref: '#/components/schemas/MenuOptionPrompt'
        - type: object
          required: [weight, is_active, created_at, updated_at]
          properties:
            weight:
              type: integer
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    StaffResponseGroup:
      allOf:
        - $ref: '#/components/schemas/ResponseGroup'
        - type: object
          required: [menu_option_id, weight, is_active, created_at, updated_at]
          properties:
            menu_option_id:
              type: integer
              format: int64
            weight:
              type: integer
              minimum: 1
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
            items:
              type: array
              items:
                $ref: '#/components/schemas/StaffResponseItem'
    StaffResponseItem:
      allOf:
        - $ref: '#/components/schemas/ResponseItem'
        - type: object
          required: [is_active, created_at, updated_at]
          properties:
            is_active:
              type: boolean
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string